Add files via upload
MK2 hardware is an attempt to make it a bit simpler/cheaper.
Uses 8 bit PWM instead of SPI DAC to drive the integrator. However, note that the range is not as large, Top midi note is now 84 (c-6). Note above this will play, but the amplitude will fall.
Also, integrator is a different design, does not need a negative supply rail.
software supports both versions of hardware, use the #define MK1hHW or MK2HW as appropriate